requirements.txt
# Core Django
Django==6.0.4
djangorestframework==3.17.1
django-redis==6.0.0
django-dirtyfields==1.9.9
django-filter==25.2

# Authentication JWT
djangorestframework_simplejwt==5.5.1

# Database
dj-database-url==3.1.2
psycopg2-binary==2.9.12

# Environment variables
python-decouple==3.8

# CORS
django-cors-headers==4.9.0

# API Documentation
drf-yasg==1.21.15

# Image handling
Pillow==12.2.0

# Celery (async tasks)
celery==5.6.3
redis==7.4.0

# Validators
phonenumbers==9.0.28
email-validator==2.3.0

# WSGI server for production
gunicorn==25.3.0
whitenoise==6.12.0
sentry-sdk==2.58.0
gevent==26.4.0

# Email
django-anymail==15.0

# Testing
coverage==7.13.5

# Utilities
Faker==40.15.0
requests==2.33.1
reportlab==4.5.0

# Media Storage (S3)
django-storages[s3]==1.14.6
boto3==1.42.93

# AI
google-genai==1.73.1
cryptography==46.0.7
